requirements.txt
# OddsIntel Engine - Python Dependencies

# Data manipulation
pandas>=2.2.0
numpy>=2.0.0

# HTTP & scraping
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
httpx>=0.27.0

# Machine learning
scikit-learn>=1.5.0
xgboost>=2.1.0
scipy>=1.14.0

# Database
supabase>=2.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Scheduling
apscheduler>=3.10.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0
click>=8.1.0

# Data validation
pydantic>=2.0.0

# Fuzzy matching
rapidfuzz>=3.0.0

# AI
google-generativeai>=0.8.0
